Moving initialize agent to chain
