+++ examples/interactive_remote_hyperdrive_example.py
@@ -34,9 +34,14 @@ private_key_1 = make_private_key()
 # Init from private key
 agent0 = chain.init_agent(
     private_key=private_key_0,
+    name="agent0",
 )
+# We can initialize an agent with a custom policy - more on that later
 agent1 = chain.init_agent(
     private_key=private_key_1,
+    name="agent1",
+    policy=PolicyZoo.random,
+    policy_config=PolicyZoo.random.Config(),
 )
 
 # %%
@@ -107,28 +112,24 @@ remove_lp_event = agent1.remove_liquidity(shares=agent1.get_lp())
 
 # Agents can also execute policies, which encapsulates actions to take on a pool.
 # This requires initializing a policy class. For example, we initialize a policy that makes random trades.
+# We can either initialize a policy on initialization (see agent1's initialization)
+# or we can explicitly call `set_policy` to set a policy on an agent.
+# NOTE: `set_policy` overwrites the existing policy.
+# TODO we may be able to set multiple policies on an agent and hot-swap them
 
-# NOTE:
-# Best practices for policies include creating a separate policy object for each agent and pool the policy
-# is expect to run on. This ensures that any internal state the policy uses is tied to a single agent and pool
-# (which most of our existing policies assume). This isn't strictly necessary for e.g., `RandomPolicy`,
-# which doesn't use state as bookkeeping (and it even may be desired to use a single policy object to e.g.,
-# use one rng state across all trades). Overall, we leave the mapping between policy objects, agents, and pools
-# to the specific policy implementation and caller.
-
-random_policy_config = PolicyZoo.random.Config(rng_seed=123)
-agent0_random_policy = PolicyZoo.random(random_policy_config)
-agent1_random_policy = PolicyZoo.random(random_policy_config)
+agent0.set_policy(
+    policy=PolicyZoo.random,
+    policy_config=PolicyZoo.random.Config(rng_seed=123),
+)
 
 # Execute policy trade on a pool
 # Output event is one of the possible trade events
 agent0_trades = []
 for i in range(10):
     # NOTE Since a policy can execute multiple trades per action, the output events is a list
-    agent0_trades.extend(agent0.execute_policy_action(policy=agent0_random_policy, pool=hyperdrive_pool))
+    agent0_trades.extend(agent0.execute_policy_action(pool=hyperdrive_pool))
 
-# Similar to pools, we can set an active policy for an agent
-agent1.set_active_policy(agent1_random_policy)
+# Agent1's policy was set during initialization
 agent1_trades = []
 for i in range(10):
     agent1_trades.extend(agent1.execute_policy_action())
